+++ src/queries/user/useMyProfileMutation/index.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import userAPI from '@/apis/users';
+import userAPI from '@/apis/user';
 import { APIUser } from '@/types/user';
 import { APIJob, APIJobGroup } from '@/types/job';
